(improvement)(headless)Restructure facade and expose only ChatLayerService and SemanticLayerService.
