+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/facade/service/ChatLayerService.java
@@ -8,6 +8,9 @@ import com.tencent.supersonic.headless.api.pojo.request.QuerySqlReq;
 import com.tencent.supersonic.headless.api.pojo.response.MapInfoResp;
 import com.tencent.supersonic.headless.api.pojo.response.MapResp;
 import com.tencent.supersonic.headless.api.pojo.response.ParseResp;
+import com.tencent.supersonic.headless.api.pojo.response.SearchResult;
+
+import java.util.List;
 
 /***dd
  * SemanticLayerService for query and search
@@ -19,6 +22,8 @@ public interface ChatLayerService {
     ParseResp performParsing(QueryNLReq queryNLReq);
 
     MapInfoResp map(QueryMapReq queryMapReq);
+
+    List<SearchResult> retrieve(QueryNLReq queryNLReq);
 
     void correct(QuerySqlReq querySqlReq, User user);
