+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/web/service/FlightService.java
@@ -1,4 +1,4 @@
-package com.tencent.supersonic.headless.server.facade.service;
+package com.tencent.supersonic.headless.server.web.service;
 
 import java.util.concurrent.ExecutorService;
 import org.apache.arrow.flight.sql.FlightSqlProducer;
